refactor(mcp): вынести DTO в подпакет dto, javadoc на каждое поле
По ревью nixel2007: рекорды из McpDtos разнесены по отдельным файлам в подпакете
mcp.dto (RangeDto, DiagnosticDto, SymbolDto, LocationDto, CallHierarchyItemDto,
DefinitionDto, CallDto); у каждого поля рекорда — javadoc (@param). Utility-класс
McpDtos удалён, импорты в инструментах обновлены.

Тесты McpToolsTest/McpHttpServerTest — зелёные; compile + spotless чистые.
