+++ src/main/java/com/github/_1c_syntax/bsl/languageserver/mcp/tools/CallHierarchyTool.java
@@ -22,8 +22,8 @@
 package com.github._1c_syntax.bsl.languageserver.mcp.tools;
 
 import com.github._1c_syntax.bsl.languageserver.mcp.McpDocumentReader;
-import com.github._1c_syntax.bsl.languageserver.mcp.McpDtos.CallDto;
-import com.github._1c_syntax.bsl.languageserver.mcp.McpDtos.CallHierarchyItemDto;
+import com.github._1c_syntax.bsl.languageserver.mcp.dto.CallDto;
+import com.github._1c_syntax.bsl.languageserver.mcp.dto.CallHierarchyItemDto;
 import com.github._1c_syntax.bsl.languageserver.providers.CallHierarchyProvider;
 import lombok.RequiredArgsConstructor;
 import org.eclipse.lsp4j.CallHierarchyIncomingCallsParams;
